Warning 7

+++ distributedcombigrid/tests/test_reduce.cpp
@@ -69,7 +69,7 @@ class TaskConst : public combigrid::Task {
     std::vector<CombiDataType>& elements = dfg_->getElementVector();
     for (auto& element : elements) {
       // BOOST_CHECK(abs(dfg_->getData()[li]));
-      element = getLevelVector()[0] / (double)getLevelVector()[1];
+      element = getLevelVector()[0] / boost::numeric_cast<double>(getLevelVector()[1]);
     }
     BOOST_CHECK(dfg_);
 
@@ -110,9 +110,9 @@ BOOST_CLASS_EXPORT(TaskConst)
 
 void checkCombine(size_t ngroup = 1, size_t nprocs = 1) {
   size_t size = ngroup * nprocs + 1;
-  BOOST_REQUIRE(TestHelper::checkNumMPIProcsAvailable(size));
+  BOOST_REQUIRE(TestHelper::checkNumMPIProcsAvailable(boost::numeric_cast<int>(size)));
 
-  CommunicatorType comm = TestHelper::getComm(size);
+  CommunicatorType comm = TestHelper::getComm(boost::numeric_cast<int>(size));
   if (comm == MPI_COMM_NULL) {
     return;
   }
@@ -124,7 +124,7 @@ void checkCombine(size_t ngroup = 1, size_t nprocs = 1) {
 
   WORLD_MANAGER_EXCLUSIVE_SECTION {
     ProcessGroupManagerContainer pgroups;
-    for (int i = 0; i < ngroup; ++i) {
+    for (int i = 0; i < boost::numeric_cast<int>(ngroup) ; ++i) {
       int pgroupRootID(i);
       pgroups.emplace_back(std::make_shared<ProcessGroupManager>(pgroupRootID));
     }
@@ -154,7 +154,7 @@ void checkCombine(size_t ngroup = 1, size_t nprocs = 1) {
     //   std::cout << coeff << std::endl;
     // }
 
-    BOOST_REQUIRE(TestHelper::checkNumMPIProcsAvailable(size));
+    BOOST_REQUIRE(TestHelper::checkNumMPIProcsAvailable(boost::numeric_cast<int>(size)));
 
     // create Tasks
     TaskContainer tasks;
@@ -167,7 +167,7 @@ void checkCombine(size_t ngroup = 1, size_t nprocs = 1) {
 
     // create combiparameters
     CombiParameters params(dim, lmin, lmax, boundary, levels, coeffs, taskIDs, ncombi);
-    params.setParallelization({nprocs, 1}); //TODO why??
+    params.setParallelization({boost::numeric_cast<long int>(nprocs), 1}); //TODO why??
 
     // create abstraction for Manager
     ProcessManager manager(pgroups, tasks, params, std::move(loadmodel));
